+++ Docs/Changes.md
@@ -14,6 +14,7 @@
   * Add: _Add `using` statement_ ([Issue #515](https://github.com/vorov2/dyalect/issues/515), [docs](https://github.com/vorov2/dyalect/wiki/Standard-operators#using)).
   * Add: _Add `FieldNotFound` error_ ([Issue #516](https://github.com/vorov2/dyalect/issues/516)).
   * Fix: _Failed test can be incorrectly displayed as successful_ ([Issue #518](https://github.com/vorov2/dyalect/issues/518)).
+  * Fix: _Comparison pattern can cause Dyalect compiler to crush_ ([Issue #519](https://github.com/vorov2/dyalect/issues/519)).
 
 # 0.25.1
   * Fix: _Exceptions thrown from overridden standard method couldn't be caught_ ([Issue #511](https://github.com/vorov2/dyalect/issues/511)).